requirements.txt
django
requests
asgiref
sqlparse
pycryptodome
